Fixes: - Pickups fall through water now (Server-side they stayed in the water surface) - Suppressed some warnings (int to short etc.) - Water is now passable for cTracer
git-svn-id: http://mc-server.googlecode.com/svn/trunk@96 0a769ca7-a7f5-676a-18bf-c427514a06d6

+++ source/cInventory.cpp
@@ -320,9 +320,9 @@ void cInventory::SendSlot( int a_SlotNum )
 	{
 		cPacket_InventorySlot InventorySlot;
 		InventorySlot.m_ItemCount = Item->m_ItemCount;
-		InventorySlot.m_ItemID = Item->m_ItemID;
-		InventorySlot.m_ItemUses = (char)Item->m_ItemHealth;
-		InventorySlot.m_SlotNum = (short)a_SlotNum;
+		InventorySlot.m_ItemID = (short) Item->m_ItemID;
+		InventorySlot.m_ItemUses = (char) Item->m_ItemHealth;
+		InventorySlot.m_SlotNum = (short) a_SlotNum;
 		InventorySlot.m_WindowID = 0; // Inventory window ID
 		m_Owner->GetClientHandle()->Send( InventorySlot );
 	}
